librbd: fix return value overflow in rbd_discard.
rbd_discard and librbd::discard both return int type value,
When discarding range is larger than 2<<31 (2G bytes), return
value will overflow.

pybind/rbd use ctypes to call librbd::rbd_discard and ctypes use
c_int as default return value type. When librbd::rbd_discard return
overflowed value, pybind/rbd will not work correctly either.

+++ src/librbd/internal.cc
@@ -2695,7 +2695,7 @@ reprotect_and_return_err:
     return mylen;
   }
 
-  int discard(ImageCtx *ictx, uint64_t off, uint64_t len)
+  ssize_t discard(ImageCtx *ictx, uint64_t off, uint64_t len)
   {
     utime_t start_time, elapsed;
     ldout(ictx->cct, 20) << "discard " << ictx << " off = " << off << " len = "
